add worker failure tests
